[BugFix] 1. Draw-Rect's position will be changed by different view mode(stereo-view or preview). 2. It works well @jj130-BT300.

+++ drawer/src/main/java/org/iii/snsi/drawer/DrawTrackingRect.java
@@ -16,7 +16,7 @@ public abstract class DrawTrackingRect extends View {
 
     public abstract void removeTrackingRect();
 
-    public abstract int[] processTrackingRect(int width, int height, int data[]);
+    public abstract int[] processTrackingRect(int width, int height, double data[]);
 
     public abstract void setLayoutSize(int width, int height);
